+++ inc/lilyParser/lilyInterpreter.ts
@@ -73,30 +73,31 @@ export class MusicTrack {
 				return [term];
 
 			const timeDenominator = context.time ? context.time.value.denominator : 4;
-			const denominator = Math.max(term.duration.denominator, timeDenominator);
+			const duration = term.durationValue;
+			const denominator = Math.max(duration.denominator, timeDenominator);
 
 			if (term.withMultiplier) {
-				const factor = term.duration.multipliers.reduce((factor, multiplier) => factor * Number(multiplier), 1);
+				const factor = duration.multipliers.reduce((factor, multiplier) => factor * Number(multiplier), 1);
 				if (!Number.isInteger(factor) || factor <= 0) {
-					console.warn("invalid multiplier:", factor, term.duration.multipliers);
+					console.warn("invalid multiplier:", factor, duration.multipliers);
 					return [term];
 				}
 
 				const event = term.clone() as MusicEvent;
 				event.duration.multipliers = [];
-	
+
 				// break duration into multiple rest events
 				const restCount = (event.duration.magnitude / WHOLE_DURATION_MAGNITUDE) * (factor - 1) * denominator;
 				if (!Number.isInteger(restCount))
 					console.warn("Rest count is not integear:", restCount, denominator, event.duration.magnitude, factor);
-	
+
 				const rests = Array(Math.floor(restCount)).fill(null).map(() =>
 					new Rest({name: "s", duration: new Duration({number: denominator, dots: 0})}));
 
 				return [event, ...rests];
 			}
 			else {
-				const divider = lcm(term.duration.subdivider, denominator);
+				const divider = lcm(duration.subdivider, denominator);
 				const restCount = term.durationMagnitude * divider / WHOLE_DURATION_MAGNITUDE;
 				console.assert(Number.isInteger(restCount), "rest count is not an integer:", restCount);